[Update] rename optim/base.py to optim/optimizer.py + refactor
